+++ src/tests/test_taktyk_receive_messages.py
@@ -1,0 +1,2 @@
+class TestTaktykReceiveMessages(object):
+    pass
